Removed code related to travel time observed from `BeamSkimmer` and extracted to `TravelTimeObserved`
