+++ src/main/scala/beam/router/TravelTimeObserved.scala
@@ -1,0 +1,229 @@
+package beam.router
+import java.awt.Color
+
+import beam.agentsim.agents.vehicles.BeamVehicleType
+import beam.agentsim.infrastructure.TAZTreeMap
+import beam.agentsim.infrastructure.TAZTreeMap.TAZ
+import beam.analysis.plots.GraphUtils
+import beam.router.Modes.BeamMode
+import beam.router.Modes.BeamMode.CAR
+import beam.sim.BeamServices
+import beam.sim.common.GeoUtils
+import beam.sim.config.BeamConfig
+import beam.utils.{FileUtils, GeoJsonReader, ProfilingUtils}
+import com.google.inject.Inject
+import com.typesafe.scalalogging.LazyLogging
+import com.vividsolutions.jts.geom.Geometry
+import org.jfree.chart.ChartFactory
+import org.jfree.chart.plot.{PlotOrientation, XYPlot}
+import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer
+import org.jfree.data.xy.{XYSeries, XYSeriesCollection}
+import org.jfree.util.ShapeUtilities
+import org.matsim.api.core.v01.{Coord, Id}
+import org.matsim.core.controler.events.IterationEndsEvent
+import org.matsim.core.utils.io.IOUtils
+import org.opengis.feature.Feature
+import org.opengis.feature.simple.SimpleFeature
+import org.supercsv.io.{CsvMapReader, ICsvMapReader}
+import org.supercsv.prefs.CsvPreference
+
+import scala.collection.mutable
+
+class TravelTimeObserved @Inject()(
+  val beamConfig: BeamConfig,
+  val beamServices: BeamServices,
+  val skimmer: BeamSkimmer
+) extends LazyLogging {
+  import TravelTimeObserved._
+
+  private val observedTravelTimesOpt: Option[Map[PathCache, Float]] = {
+    val zoneBoundariesFilePath = beamConfig.beam.calibration.roadNetwork.travelTimes.zoneBoundariesFilePath
+    val zoneODTravelTimesFilePath = beamConfig.beam.calibration.roadNetwork.travelTimes.zoneODTravelTimesFilePath
+
+    if (zoneBoundariesFilePath.nonEmpty && zoneODTravelTimesFilePath.nonEmpty) {
+      val tazToMovId: Map[TAZ, Int] = buildTAZ2MovementId(
+        zoneBoundariesFilePath,
+        beamServices.geo,
+        beamServices.tazTreeMap
+      )
+      val movId2Taz: Map[Int, TAZ] = tazToMovId.map { case (k, v) => v -> k }
+      Some(buildPathCache2TravelTime(zoneODTravelTimesFilePath, movId2Taz))
+    } else None
+  }
+
+  val uniqueModes: List[BeamMode.CAR.type] = List(CAR)
+
+  val uniqueTimeBins: Range.Inclusive = 0 to 23
+
+  val dummyId: Id[BeamVehicleType] = Id.create("NA", classOf[BeamVehicleType])
+
+  def writeTravelTimeObservedVsSimulated(event: IterationEndsEvent): Unit = {
+    observedTravelTimesOpt.foreach { observedTravelTimes =>
+      ProfilingUtils.timed(
+        s"writeTravelTimeObservedVsSimulated on iteration ${event.getIteration}",
+        x => logger.info(x)
+      ) {
+        write(event, observedTravelTimes)
+      }
+    }
+  }
+
+  private def write(event: IterationEndsEvent, observedTravelTimes: Map[PathCache, Float]): Unit = {
+    val filePathObservedVsSimulated = event.getServices.getControlerIO.getIterationFilename(
+      event.getServices.getIterationNumber,
+      "tazODTravelTimeObservedVsSimulated.csv.gz"
+    )
+    val writerObservedVsSimulated = IOUtils.getBufferedWriter(filePathObservedVsSimulated)
+    writerObservedVsSimulated.write("fromTAZId,toTAZId,hour,timeSimulated,timeObserved")
+    writerObservedVsSimulated.write("\n")
+
+    val series: XYSeries = new XYSeries("Time", false)
+
+    beamServices.tazTreeMap.getTAZs
+      .foreach { origin =>
+        beamServices.tazTreeMap.getTAZs.foreach { destination =>
+          uniqueModes.foreach { mode =>
+            uniqueTimeBins
+              .foreach { timeBin =>
+                val key = PathCache(origin.tazId, destination.tazId, timeBin)
+                observedTravelTimes.get(key).foreach { timeObserved =>
+                  val theSkim = skimmer
+                    .getSkimValue(timeBin * 3600, mode, origin.tazId, destination.tazId)
+                    .map(_.toSkimExternal)
+                    .getOrElse {
+                      if (origin.equals(destination)) {
+                        val newDestCoord = new Coord(
+                          origin.coord.getX,
+                          origin.coord.getY + Math.sqrt(origin.areaInSquareMeters) / 2.0
+                        )
+                        skimmer.getSkimDefaultValue(
+                          mode,
+                          origin.coord,
+                          newDestCoord,
+                          timeBin * 3600,
+                          dummyId,
+                          beamServices
+                        )
+                      } else {
+                        skimmer.getSkimDefaultValue(
+                          mode,
+                          origin.coord,
+                          destination.coord,
+                          timeBin * 3600,
+                          dummyId,
+                          beamServices
+                        )
+                      }
+                    }
+
+                  series.add(theSkim.time, timeObserved)
+                  writerObservedVsSimulated.write(
+                    s"${origin.tazId},${destination.tazId},${timeBin},${theSkim.time},${timeObserved}\n"
+                  )
+                }
+              }
+          }
+        }
+      }
+
+    writerObservedVsSimulated.close()
+
+    val chartPath =
+      event.getServices.getControlerIO.getIterationFilename(event.getServices.getIterationNumber, chartName)
+    generateChart(series, chartPath)
+  }
+}
+
+object TravelTimeObserved extends LazyLogging {
+  val chartName: String = "scatterplot_simulation_vs_reference.png"
+
+  case class PathCache(from: Id[TAZ], to: Id[TAZ], hod: Int)
+
+  def buildTAZ2MovementId(filePath: String, geo: GeoUtils, tazTreeMap: TAZTreeMap): Map[TAZ, Int] = {
+    ProfilingUtils.timed(s"buildTAZ2MovementId from '$filePath'", x => logger.info(x)) {
+      val mapper: Feature => (TAZ, Int, Double) = (feature: Feature) => {
+        val centroid = feature.asInstanceOf[SimpleFeature].getDefaultGeometry.asInstanceOf[Geometry].getCentroid
+        val wgsCoord = new Coord(centroid.getX, centroid.getY)
+        val utmCoord = geo.wgs2Utm(wgsCoord)
+        val movId = feature.getProperty("MOVEMENT_ID").getValue.toString.toInt
+        val taz: TAZ = tazTreeMap.getTAZ(utmCoord.getX, utmCoord.getY)
+        val distance = geo.distUTMInMeters(utmCoord, taz.coord)
+        (taz, movId, distance)
+      }
+      val xs: Array[(TAZ, Int, Double)] = GeoJsonReader.read(filePath, mapper)
+      val tazId2MovIdByMinDistance = xs
+        .groupBy { case (taz, _, _) => taz }
+        .map {
+          case (taz, arr) =>
+            val (_, movId, _) = arr.minBy { case (_, _, distance) => distance }
+            (taz, movId)
+        }
+      val end = System.currentTimeMillis()
+      val numOfUniqueMovId = xs.map(_._2).distinct.size
+      logger.info(
+        s"xs size is ${xs.size}. tazId2MovIdByMinDistance size is ${tazId2MovIdByMinDistance.keys.size}. numOfUniqueMovId: $numOfUniqueMovId"
+      )
+      tazId2MovIdByMinDistance
+    }
+  }
+
+  def buildPathCache2TravelTime(pathToAggregateFile: String, movId2Taz: Map[Int, TAZ]): Map[PathCache, Float] = {
+    val observedTravelTimes: mutable.HashMap[PathCache, Float] = scala.collection.mutable.HashMap.empty
+    ProfilingUtils.timed(s"buildPathCache2TravelTime from '$pathToAggregateFile'", x => logger.info(x)) {
+      val mapReader: ICsvMapReader =
+        new CsvMapReader(FileUtils.readerFromFile(pathToAggregateFile), CsvPreference.STANDARD_PREFERENCE)
+      try {
+        val header = mapReader.getHeader(true)
+        var line: java.util.Map[String, String] = mapReader.read(header: _*)
+        while (null != line) {
+          val sourceid = line.get("sourceid").toInt
+          val dstid = line.get("dstid").toInt
+          val mean_travel_time = line.get("mean_travel_time").toFloat
+          val hod = line.get("hod").toInt
+
+          if (movId2Taz.contains(sourceid) && movId2Taz.contains(dstid)) {
+            observedTravelTimes.put(PathCache(movId2Taz(sourceid).tazId, movId2Taz(dstid).tazId, hod), mean_travel_time)
+          }
+
+          line = mapReader.read(header: _*)
+        }
+      } finally {
+        if (null != mapReader)
+          mapReader.close()
+      }
+    }
+    logger.info(s"observedTravelTimesOpt size is ${observedTravelTimes.keys.size}")
+    observedTravelTimes.toMap
+  }
+
+  def generateChart(series: XYSeries, path: String): Unit = {
+    val dataset = new XYSeriesCollection
+    dataset.addSeries(series)
+    val chart = ChartFactory.createScatterPlot(
+      "TAZ TravelTimes Observed Vs. Simulated",
+      "Simulated",
+      "Observed",
+      dataset,
+      PlotOrientation.VERTICAL,
+      false,
+      true,
+      false
+    )
+
+    val xyplot = chart.getPlot.asInstanceOf[XYPlot]
+
+    val renderer = new XYLineAndShapeRenderer
+    renderer.setSeriesShape(0, ShapeUtilities.createDiamond(1))
+    renderer.setSeriesPaint(0, Color.RED)
+    renderer.setSeriesLinesVisible(0, false)
+
+    xyplot.setRenderer(0, renderer)
+
+    GraphUtils.saveJFreeChartAsPNG(
+      chart,
+      path,
+      1000,
+      1000
+    )
+  }
+}
