+++ src/main/scala/beam/sim/BeamSim.scala
@@ -14,7 +14,7 @@ import beam.analysis.plots.{GraphUtils, GraphsStatsAgentSimEventsListener}
 import beam.analysis.via.ExpectedMaxUtilityHeatMap
 import beam.analysis.{DelayMetricAnalysis, IterationStatsProvider}
 import beam.physsim.jdeqsim.AgentSimToPhysSimPlanConverter
-import beam.router.{BeamRouter, BeamSkimmer, RouteHistory}
+import beam.router.{BeamRouter, BeamSkimmer, RouteHistory, TravelTimeObserved}
 import beam.router.gtfs.FareCalculator
 import beam.router.osm.TollCalculator
 import beam.sim.metrics.MetricsPrinter.{Print, Subscribe}
@@ -61,7 +61,8 @@ class BeamSim @Inject()(
   private val scenario: Scenario,
   private val networkHelper: NetworkHelper,
   private val beamOutputDataDescriptionGenerator: BeamOutputDataDescriptionGenerator,
-  private val beamSkimmer: BeamSkimmer
+  private val beamSkimmer: BeamSkimmer,
+  private val travelTimeObserved: TravelTimeObserved
 ) extends StartupListener
     with IterationStartsListener
     with IterationEndsListener
@@ -168,6 +169,11 @@ class BeamSim @Inject()(
   }
 
   override def notifyIterationEnds(event: IterationEndsEvent): Unit = {
+    // ORDER IS IMPORTANT HERE!
+    // travelTimeObserved use skimmer to write `TravelTimeObservedVsSimulated`
+    // `beamSkimmer.notifyIterationEnds(event)` will clear skims, so `travelTimeObserved.writeTravelTimeObservedVsSimulated` must be first
+    travelTimeObserved.writeTravelTimeObservedVsSimulated(event)
+
     beamSkimmer.notifyIterationEnds(event)
 
     if (beamServices.beamConfig.beam.debug.debugEnabled)
